+++ plugins/scaffolder-actions/scaffolder-backend-module-utils/src/actions/jsonata/json.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Larder Software Limited
+ * Copyright 2023 Larder Software Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import { getVoidLogger } from '@backstage/backend-common';
 import { createJsonJSONataTransformAction } from './json';
 import { PassThrough } from 'stream';
@@ -34,7 +35,7 @@ describe('roadiehq:utils:jsonata:json:transform', () => {
   };
   const action = createJsonJSONataTransformAction();
 
-  it('should write file to the workspacePath with the given transformation', async () => {
+  it('should output default string result of having applied the given transformation', async () => {
     mock({
       'fake-tmp-dir': {
         'fake-file.json': '{ "hello": ["world"] }',
@@ -53,5 +54,48 @@ describe('roadiehq:utils:jsonata:json:transform', () => {
       'result',
       JSON.stringify({ hello: ['world', 'item2'] }),
     );
+  });
+
+  it('should output string result of having applied the given transformation', async () => {
+    mock({
+      'fake-tmp-dir': {
+        'fake-file.json': '{ "hello": ["world"] }',
+      },
+    });
+    await action.handler({
+      ...mockContext,
+      workspacePath: 'fake-tmp-dir',
+      input: {
+        path: 'fake-file.json',
+        expression: '$ ~> | $ | { "hello": [hello, "item2"] }|',
+        as: 'string',
+      },
+    });
+
+    expect(mockContext.output).toHaveBeenCalledWith(
+      'result',
+      JSON.stringify({ hello: ['world', 'item2'] }),
+    );
+  });
+
+  it('should output object result of having applied the given transformation', async () => {
+    mock({
+      'fake-tmp-dir': {
+        'fake-file.json': '{ "hello": ["world"] }',
+      },
+    });
+    await action.handler({
+      ...mockContext,
+      workspacePath: 'fake-tmp-dir',
+      input: {
+        path: 'fake-file.json',
+        expression: '$ ~> | $ | { "hello": [hello, "item2"] }|',
+        as: 'object',
+      },
+    });
+
+    expect(mockContext.output).toHaveBeenCalledWith('result', {
+      hello: ['world', 'item2'],
+    });
   });
 });