+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,56 @@
 # backend
 
+## 2.0.17
+
+### Patch Changes
+
+- Updated dependencies [05c0c417]
+  - @roadiehq/backstage-plugin-aws-backend@1.1.12
+  - @roadiehq/backstage-plugin-argo-cd-backend@2.8.2
+  - @roadiehq/backstage-plugin-aws-auth@0.4.12
+  - @roadiehq/catalog-backend-module-aws@1.3.13
+  - @roadiehq/catalog-backend-module-okta@0.8.5
+  - @roadiehq/scaffolder-backend-module-aws@2.4.11
+  - @roadiehq/scaffolder-backend-module-http-request@4.0.10
+  - @roadiehq/scaffolder-backend-module-utils@1.10.1
+
+## 2.0.16
+
+### Patch Changes
+
+- Updated dependencies [edf2f0ab]
+  - @roadiehq/backstage-plugin-aws-backend@1.1.11
+  - @roadiehq/backstage-plugin-argo-cd-backend@2.8.1
+  - @roadiehq/backstage-plugin-aws-auth@0.4.11
+  - @roadiehq/catalog-backend-module-aws@1.3.12
+  - @roadiehq/catalog-backend-module-okta@0.8.4
+  - @roadiehq/scaffolder-backend-module-aws@2.4.10
+  - @roadiehq/scaffolder-backend-module-http-request@4.0.9
+  - @roadiehq/scaffolder-backend-module-utils@1.8.9
+
+## 2.0.15
+
+### Patch Changes
+
+- Updated dependencies [18f8a82f]
+  - @roadiehq/backstage-plugin-argo-cd-backend@2.8.0
+
+## 2.0.14
+
+### Patch Changes
+
+- 608e1061: Release all
+- Updated dependencies [608e1061]
+  - app@1.0.10
+  - @roadiehq/backstage-plugin-aws-backend@1.1.10
+  - @roadiehq/backstage-plugin-argo-cd-backend@2.7.8
+  - @roadiehq/backstage-plugin-aws-auth@0.4.10
+  - @roadiehq/catalog-backend-module-aws@1.3.11
+  - @roadiehq/catalog-backend-module-okta@0.8.3
+  - @roadiehq/scaffolder-backend-module-aws@2.4.9
+  - @roadiehq/scaffolder-backend-module-http-request@4.0.8
+  - @roadiehq/scaffolder-backend-module-utils@1.8.8
+
 ## 2.0.13
 
 ### Patch Changes