+++ plugins/backend/catalog-backend-module-aws/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-catalog-backend-module-aws
 
+## 1.3.13
+
+### Patch Changes
+
+- 05c0c417: Update deps to backstage 1.16.0 version
+
+## 1.3.12
+
+### Patch Changes
+
+- edf2f0ab: Update dependencies to backstage version 1.15.0
+
+## 1.3.11
+
+### Patch Changes
+
+- 608e1061: Release all
+
 ## 1.3.10
 
 ### Patch Changes