+++ plugins/backend/catalog-backend-module-okta/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @roadiehq/catalog-backend-module-okta
 
+## 0.8.5
+
+### Patch Changes
+
+- 05c0c417: Update deps to backstage 1.16.0 version
+
+## 0.8.4
+
+### Patch Changes
+
+- edf2f0ab: Update dependencies to backstage version 1.15.0
+
+## 0.8.3
+
+### Patch Changes
+
+- 608e1061: Release all
+
 ## 0.8.2
 
 ### Patch Changes