+++ plugins/home/backstage-plugin-home-markdown/CHANGELOG.md
@@ -1,5 +1,23 @@
 ## 1.1.0
 
+## 2.5.13
+
+### Patch Changes
+
+- 05c0c417: Update deps to backstage 1.16.0 version
+
+## 2.5.12
+
+### Patch Changes
+
+- edf2f0ab: Update dependencies to backstage version 1.15.0
+
+## 2.5.11
+
+### Patch Changes
+
+- 608e1061: Release all
+
 ## 2.5.10
 
 ### Patch Changes