+++ utils/roadie-backstage-entity-validator/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @roadiehq/roadie-backstage-entity-validator
 
+## 2.3.5
+
+### Patch Changes
+
+- 05c0c417: Update deps to backstage 1.16.0 version
+
+## 2.3.4
+
+### Patch Changes
+
+- f7e24d24: Allow for backstage default mkdocs.yaml
+
+## 2.3.3
+
+### Patch Changes
+
+- edf2f0ab: Update dependencies to backstage version 1.15.0
+
+## 2.3.2
+
+### Patch Changes
+
+- 608e1061: Release all
+
 ## 2.3.1
 
 ### Patch Changes